// backend/src/common/errors/error-codes.enum.ts
export enum ErrorCode {
	INVALID_INPUT = "INVALID_INPUT",
	DECK_NOT_FOUND = "DECK_NOT_FOUND",
	INVALID_CREDENTIALS = "INVALID_CREDENTIALS",
	USER_NOT_FOUND = "USER_NOT_FOUND",
	EMAIL_ALREADY_EXISTS = "EMAIL_ALREADY_EXISTS",
	CPF_OR_CNPJ_ALREADY_EXISTS = "CPF_OR_CNPJ_ALREADY_EXISTS",
	PHONE_NUMBER_ALREADY_EXISTS = "PHONE_NUMBER_ALREADY_EXISTS",
	DATABASE_ERROR = "DATABASE_ERROR",
	MISSING_ACCEPT_TERMS = "MISSING_ACCEPT_TERMS",
	INVALID_TOKEN = "INVALID_TOKEN",
	TOKEN_EXPIRED = "TOKEN_EXPIRED",
	INVALID_TERMS_ID = "INVALID_TERMS_ID",
	USER_ALREADY_ACCEPTED_TERMS = "USER_ALREADY_ACCEPTED_TERMS",
	LOGIN_FAILED = "LOGIN_FAILED",
	UNAUTHORIZED_PERMISSION = "UNAUTHORIZED_PERMISSION",
	VERSION_ALREDY_EXIST = "VERSION_ALREDY_EXIST",
	NO_ACTIVE_TERMS_FOUND = "NO_ACTIVE_TERMS_FOUND",
	TERMS_NOT_FOUND = "TERMS_NOT_FOUND",
}